Add files via upload
In "work.sql," you will create all the necessary data to implement a situation of retail, where you have the information of three stores in different parts of USA and the respective sales of all the stores between the years 2016 and 2018. With this information, you need to create the schemas and load the data to filtred all the possible information and give it to the business section of the company. 
Then with "Query_work.sql," present the Query realized in MySQL to obtain all the primary information of these three stores, present it to the company's business section, and create a little report.